requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
